Changes to plot stucture for a presentation and general progress on updating some of the scritps.
Plots and figures altered to better suit a presentation format, as well as updates made to allow four polts (one for each gridspacing of 5, 15, 22.5, and 30 degrees). Violin plot should now be functional as well (updated to python 3). Some lines in script commented out to produce the new versions of plots (such as the change to a scatter plot from a bar plot to better show some of the data).  The commented lines can be used to revert to the previous plotting formats.